Reorganize everything under src into lib, components, etc
