+++ src/components/partbuilder/parts-list-cell.vue
