+++ src/components/partbuilder/parts-list-footer.vue
