+++ src/components/shipdesigner/module.vue
@@ -42,9 +42,9 @@
 import {
 	pretty,
 	frac,
-} from './ui-functions.js';
+} from '@/lib/ui-functions';
 
-import StatlineCell from './statline-cell.vue';
+import StatlineCell from '@/components/shipdesigner/statline-cell.vue';
 
 export default {
 	name: 'ModuleTr',
