+++ src/components/shipdesigner/setting-cell.vue
