+++ src/components/shipdesigner/statline-cell.vue
