+++ src/components/shipdesigner/subsystem-frame.vue
@@ -36,9 +36,9 @@
 
 <script>
 
-import StatlineCell from './statline-cell.vue';
+import StatlineCell from '@/components/shipdesigner/statline-cell.vue';
 
-import { frac } from './ui-functions.js';
+import { frac } from '@/lib/ui-functions';
 
 export default {
 	name: 'SubsystemFrame',
