+++ src/components/shipdesigner/subsystem-summary.vue
@@ -34,9 +34,9 @@
 
 <script>
 
-import StatlineCell from './statline-cell.vue';
+import StatlineCell from '@/components/shipdesigner/statline-cell.vue';
 
-import { frac } from './ui-functions.js';
+import { frac } from '@/lib/ui-functions.js';
 
 export default {
 	name: 'SubsystemSummary',
