+++ src/lib/namedvector.js
