+++ src/lib/shipimporter.js
